naming-and-structure
